Change the file content function
